fix(docs): docs build in aarch64 fails
docs build in aarch64 fails as PythonModules.cmake is not compatible
with cmake version 3.16.3. Checking for the presence of python modules
is not required as the python modules found are not directly used by
the cmake rule.

This fix also removes FindPlantUML. Dependency on PlantUML was removed
as part of commit f1a13aa.

Git release tag will be in format tf-rmm-<version>. Add tf-rmm- prefix
in docs/conf.py

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: Ife9dc13205abb145c723354cc5843cfd436bf5ef