fix(build): Workaround warning during MbedTLS configuration
The following message is output to the terminal when MbedTLS
is configured.

GNUInstallDirs.cmake:233 (message):
  Unable to determine default CMAKE_INSTALL_LIBDIR directory because no
  target architecture is known.  Please enable at least one language before
  including GNUInstallDirs.
Call Stack (most recent call first):
  ext/mbedtls/CMakeLists.txt:42 (include)
This warning is for project developers.  Use -Wno-dev to suppress it.

This is because the CMake variable CMAKE_SIZEOF_VOID_P is not set
correctly when using AArch64 toolchain. This variable is used by
mbedTLS lib install step to figure out the right path to install
libs for a Multi-arch host. This is not really used when MbedTLS is
used with RMM and hence we hardcode this value to `8` when RMM_ARCH
is aarch64.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: Ic066c61c2707c7df45f9763469136bcd415819f1