fix(granules): Properly check granule alignment on granule_addr()
The commit aecad126cf873639 introduced a check to ensure that
`struct granule *` parameter is aligned to the sizeof struct
granule. But 'granules' array might not be aligned to struct
granule size and hence introduced a regression. This is now
fixed and the assertion check in granule_addr() takes this
into account.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: Ie75f099c1423140f7e4828603ecd92692fc2452d
Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>