fix(lib/realm): remove assert() to allow GCC 13 compile
The recent RSI/RMI header restructing introduced an assertion check in
status_ptr() function. This was causing an issue with GCC 13 with the
following error:

~/tf-rmm/runtime/rmi/realm.c
In file included from ~/tf-rmm/runtime/rmi/realm.c:9:
granule.h: In function 'find_lock_unused_granule':
granule.h:224:42: error: edge points to wrong declaration:
  224 |  return (struct granule *)status_ptr(RMI_ERROR_INPUT);
      |                           ^~~~~~~~~~~~~~~~~~~~~~~~~~~
 <function_decl 0x7fdfc8eb0900 __builtin_trap
    type <function_type 0x7fdfc8deb5e8

There was no obvious issue in the code and removing the assert()
in status_ptr() resolved the issue. Given that the assert() was not
providing any significant benefit to the functionality, the decision was
made to remove the assert and resolve the build error.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I7a419491ccea235c8dee4f11bdb32eaf90b3b947