refactor(attest): use PSA Crypto API
This commit changes the attestation implementation in RMM to use PSA
crypto API instead of the legacy MbedTLS API.

RMM uses a restartable implementation of the signing operation in
t_cose. It is important to note that this interface is not merged to
t_cose but exists as a PR. This commit changes the t_cose submodule to
the necessary commit in a forked t_cose repo.
The PR is built upon the dev branch in t_cose. This branch require a
more recent version of qcbor, so that submodule is updated as well.
The t_cose implementation mentioned above uses a feature from MbedTLS
that is released in v3.4.0. So the hash of the Mbed TLS submodule is
updated as well.

One important detail is that the random number generation
(lib/attestation/src/attestation_rnd.c) still uses the legacy Mbed TLS
APIs for pseudo random number generation, as this is a function not
supported by the PSA Crypto API.

This commit also adds patches to Mbed TLS as the version used in this
commit fails to build with clang using RMM's config. For details see
commit messages in the patches.

Change-Id: I1ccabd096572316fd8260a47d709b0f642c0332d
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>