fix(rmm): remove HVC exit handling from REC_ENTER handler
The older version of the RMM used to transfer HVC calls to the
host. That has been since replaced by the RSI_HOST_CALL but
the code that handles the REC exit due to HVC wrongly remained
in REC_ENTER handler. Hence remove the HVC handling code
in REC_ENTER.

Signed-off-by: Alexei Fedorov <Alexei.Fedorov@arm.com>
Change-Id: I6775f8062f9bc3e5c41aefd837b34dc5ede0f390