feat(fake_host):  Enable generating profile data
This patch adds necessary CFlags for generating the profiling
data when binary is executed. A grof file `gmon.out` is generated
when the fake_host binary is executed and it can be processed
using the gprof tool. A sample gprof cmd is shown below:

> gprof build/Debug/rmm.elf > prof.txt

The `fomit-frame-pointer` was not compatible with profiling options
hence this patch moves the option to make it AArch64 specific.
Also `-gdwarf-4` option is made common for fake_host and
AArch64.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: Id988e7fcb18048cc4d34ced3a339e73b116be786