fix(fvp): Increase the RMM_MAX_GRANULES
Increase RMM_MAX_GRANULES so that RMM can cater for the case
that the linux DT nodes parsed by TF-A has not accounted
for the 64 MB carveout. It is possible to mark the carveout
in DT by other mechanisms (like `memreserve` property) and
TF-A is not taking these mechanisms into account before
communicating the DDR memory range to RMM. This results
in an additional 64MB than RMM_MAX_GRANULES value. This patch
increases the RMM_MAX_GRANULES to cater for this 64 MB as well
as there is no impact to RMM other than slightly higher
memory footprint.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I8148b4d48e40b17586d6d8f942328d4842bc28c6