feat(xlat): add unittests for FEAT_LPA2 support on stage 1 xlat
This patch fixes the current unittests for the stage 1
translation table to support FEAT_LPA2 as well as it adds
some specific tests for that feature.

It also fixes all the warning and errors on the .cpp files
for the xlat library as reported by checkpatch.pl and adds
some missing tests.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I910c14bd6f4f3bfa1ebbf165d9693f6442e91fb2