refactor(rmm_el3_ifc): harden compatibility checks for version
This patch hardens the requirements for compatibility for RMM-EL3
version. The major version has to match and minor version needs to
be >= the number expected by RMM. This RMM_EL3_IFC version is bumped
up to 0.2 to reflect the change in expectation wrt to save and restore
of PAuth registers.

New helper compatibility macros are introduced to ease the version
checks in code. The `plat_dram` feild check in manifest parsing is
also hardened to expect NULL.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: Id9fb8e727cb4084fdf8ba3fcb191a15e3575088c