build(mbedtls): harden patch application
The RMM repo applies patch to MbedTLS to resolve some
build errors. Previously the patch application was skipped
if the git diff resulted in some output. This patch ensures
that MbedTLS is checked out and the patch is successfully
applied for every build.

Note: Be aware that any local changes done to MbedTLS will
be reverted as part of the build process. Ensure to check-in
any changes if they are required to be included in the build.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I0d174dcecccecdf1362990fbb12efadce6fe62b5