fix(lib/arch): add function to clear SVE P and FFR registers
In SIMD context switch, while restoring SVE context explicitly clear
SVE P and FFR registers if SVE hint was set. If not cleared the outgoing
context might have values in P and FFR registers of a the last used
SIMD context.

This is not required for Z registers as restoring the Q[0-31] registers
will clear all the upper bits of Z[0-31] registers.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: Ia0c8f7341f33998a91aa918aab35e07ae075fffa