fix(tools): add file exclusions in checkpatch
Currently, certain directories and files are excluded from checkpatch
analysis when running on the codebase (target `checkcodebase`). However,
these exclusions are not being applied when running on pending commits
(target `checkpatch`).

This patch modifies the `checkpatch` target so that any file exclusions
which apply to `checkcodebase` will also apply to `checkpatch`.

Signed-off-by: Chuyue Luo <Chuyue.Luo@arm.com>
Change-Id: I8236363def250d3ed44bc94ba14324c740a8fd23