fix(lib/qcbor): correct linker options
The QCBOR cmake was setting the `-lm` option to linker options.
Although this was ignored in previous versions of GCC, the
aarch64-none-linux-gnu 12 picks this up and tries to link RMM
with math lib.

This patch fixes the issue by setting the right options to QCBOR
as part of configuration which then allows the QCBOR to do the
right checks and not set the `-lm` linker option.

Fixes TF-RMM/tf-rmm#11

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I8e5a9551838061b281ff3a6ac3c50cd1a2fafe50