fix(lib/xlat): Align structures accessed by secondary CPUs
There are a series of data structures to be initialized by secondary
CPUs before they can enable MMU and consequently access Normal cacheable
memory. There was a bug that if these structures are not cache-line
aligned, then there is a chance that they may be placed adjacent
to a data which is accessed when MMU is ON thus making the cache
line dirty. This has the consequence that when the xlat structures
are initialized by secondary CPUs with MMU off, there is a chance
that the dirty line in cache will be written back to memory, thus
corrupting the update done by secondary.

This patch fixes that bug by ensuring that xlat structures are cache
line aligned. Also, the previous flush_dcache_range() calls are
replaced with inv_dcache_range() when data is modified with MMU is
OFF as this is the more correct operation to be performed.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: Id98daf28fbe5fa636d1a3e3466ab7ed2fbc630c5