fix(lib/xlat): Assertions for `is_mmu_enabled`.
The runtime checks for some of the xlat APIs for
is_mmu_enabled() are replaced with assertions so as to make it
consistent across the RMM codebase.

Also, added assertions to some APIs to ensure that the
preconditions for calling the API are adhered to.

Modified the unit tests to check the above added assertion.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I03aecef79a8dd259a0755808b16f0218d44f4180