fix(rmm): initialize MDCR_EL2.HPMN for each REC
MDCR_EL2.HPMN defines the number of PMU counters available to a Realm.
Writing the MDCR_EL2 register from validate_realm_params() doesn't apply
the change to the Realm being configured, and only happens on the
current CPU. Instead, update the value of MDCR_EL2 for each REC, so that
it gets applied whenever the REC gets scheduled.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Change-Id: I01009648f94f60f51824ab2c25d93f00d80b984a