fix(tools): merge checkpatch and commit message checker counts
Currently, the error and warning counts for the checkpatch.pl and
checkcommitmsg.py scripts are outputted separately. An issue with this
is that, if there is an error in checkpatch.pl, CheckPatch.cmake will
exit early and the counts for checkcommitmsg.py will not be printed.

This patch updates CheckPatch.cmake to print the total error and warning
counts from these two scripts, rather than printing the counts from each
script separately. In addition, this patch updates checkcommitmsg.py to
count the number of warnings. Finally, the error and warning messages
for the commit message checker have been modified to make it clear that
they apply to the commit message.

Signed-off-by: Chuyue Luo <Chuyue.Luo@arm.com>
Change-Id: Ie6855192ac01e81eb43ed4ea7d224cf0583a678b