fix(tools): fix issue with CheckPatch output being truncated
There is a bug in CheckPatch.cmake where the output of checkpatch.pl is
truncated at the character ';', if it is present. This is because the
output is treated as a list by CMake. The fix is to enclose these
variables with double quote marks so they will be treated as text.

Signed-off-by: Chuyue Luo <Chuyue.Luo@arm.com>
Change-Id: I00e309f10b14bdafd56b21d2812dce8a7299808a