fix(plat/host): Fix CBMC errors related to DCZVA patch
The patch 862f96c introduced support for DCZVA but it also introduced
some errors for CBMC analysis. Since the GRANULE_SHIFT is 7 for CBMC
builds, the value of 9 for DCZID_EL0_BS in host platform caused overflow
for one of the calculations. This patch reduces the value for host
platform to 5.

The `cbmc-gcc` build target is also fixed by this patch.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I210d54723f36b1aae07caffc3bfa986cf2341e55