build(lib): Add extra repositories for system includes
Toolchains such as aarch64-none-elf, aarch64-none-linux-gnu and
aarch64-linux-gnu include assert.h and limits.h in a directory that is
part of their search path.  This is not the case when compiling with
Yocto where aarch64-poky-linux places those files in the sysroot
directory of the component being compiled.

Since the sysroot directory of the component is not part of the cmake
search path, compiling the RMM in Yocto fails.  This patch fixes the
problem by expanding the search path when needed, allowing the RMM to be
compiled in Yocto.

Change-Id: I0393d7f6fb26b31292515251628ab8a74ce0cbda
Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>