fix(toolchains/aarch64): replace ZIP_LISTS with string REPLACE
When trying to search for the highest architecture revision supported by
the compiler, we need to use a string in the form of "8.5" for the
compiler option, but something like "8_5" for the internal variables.
At the moment this is solved with the help of cmake's ZIP_LISTS feature.
However this was only introduced in cmake 3.17, but our recommended
build platform (Ubuntu 20.04) ships with cmake 3.16.3 only.

Replace the usage of ZIP_LISTS with a simple "string(REPLACE" command,
that is compatible with much older cmake versions.

This allows to build RMM with Ubuntu's stock cmake again.

Change-Id: I24df30228ed3100f1508af84dfc8e6b577693f64
Signed-off-by: Andre Przywara <andre.przywara@arm.com>