fix(tools/cppcheck): suppress misra warnings
Suppress misra warning rules-
misra-c2012-12.3 - Comma should not be used
misra-config - To suppress COMPILER_ASSERT warnings

Suppress cppcheck warning due to const params/variables.
Also added `checkers-report` to the cppcheck cmdline as
this gives details on the checkers invoked and there
was an error from the tool without this option.

Change-Id: I56701f11794307a854521e087455811d01bc9577
Signed-off-by: Shruti Gupta <shruti.gupta@arm.com>