feat(tools/shrinkwrap): add shrinkwrap configs for RMM
This patch adds a shrinkwrap configuration to build and run RMM TFTF
tests using the current local (or, alternatively, a different one)
RMM repository. The configuration included in this patch can be
found at:

	- <RMM_ROOT>/shrinkwrap/configs/rmm-tftf.yaml

In addition to the above configuration, this patch includes the
following overlays, that can be used in any needed combination:

	- rmm.yaml
	- rmm-debug.yaml
	- model-enable-lpa2.yaml
	- model-wait-debugger.yaml
	- clean.yaml

Of all the above overlays, which are stored on the same directory as
the configuration, the first one is already used by the former
whereas the rest of overlays can be used as needed to test any
particular functionality.

In addition to the presented configuration and overlays, this patch
also adds instructions to the Getting Started guide to build
and run the 3-world CCA demonstrator. This is done by using
the `rmm.yaml` overlay provided in the patch,
so it can be used to test RMM during development.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I8ccd243d8580ad6ddb4b146674f55c8d6b5b1df9