feat(mec): Add MECID programming
For RTT and DATA granules, RMM should program the Alternate MECID
(mecid_a1_el2) register before mapping the same to Stage1 and
accessing it. Similarly, when the Realm is executing, the VMEC
Primary MECID (vmecid_p_el2) register should be programmed.

A new helper function map_rd_and_init_realm_mecid_s1() is
introduced which maps RD and also programs the mecid_a1_el2
register. This avoids the need to explicitly program this register
before a mapping to RTT and DATA granule. Also, assertion is
added to buffer_map_internal() to ensure that this register is
programmed before the granules are mapped to Stage 1 of RMM.

This helper is intended to be called by most RMI calls.
RMI_REC_CREATE() and RMI_REC_AUX_COUNT() does not make use of the
helper as it does not need any pages to be mapped with Alternate
MECID.

A call to mec_init_realm_mecid_s2() is added which will
program the vmecid_p_el2 register prior to entering the Realm.

The registers are reset as part of RMI exit sequence in handler.c.

The patch also adds a check for RMM_V1_1 in is_feat_mec_present()
call since MEC features are not meant to be present in a RMM v1.0
implementation.

Change-Id: Ib80700ec94da01875cfa7cba0b09146566805dc4
Signed-off-by: Tushar Khandelwal <tushar.khandelwal@arm.com>
Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>