refactor(lib): restore COMPILER_ASSERT for CBMC
COMPILER_ASSERT is a useful tool for CBMC to detect issues that can
lead to assert failures. A previous commit completely turned off
COMPILER_ASSERT because some alignment and structure size requirements
are not fulfilled by the CBMC configuration.

This commit introduces a new macro COMPILER_ASSERT_NO_CBMC to be used
when an assertion is only valid in regular build, but not in CBMC build.

Change-Id: I20cd80bd5ca86ee92407c92e26ac2c99698a3e05
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>