fix(lib): Adjust heap size based on MAX_CPUS
The PRNG portion of the heap size for the mbedtls library varies based
on MAX_CPUS but the portion reserved for other allocation is statically
set to 4K, leading to a situation where the RMM runs out of memory when
the Realm attestation key is retrieved and processed.

This patch set the portion reserved to other allocations in accordance
with the number of CPUS available in the system.

Change-Id: I7960098d2f9af3f1b8e8d0890284253f99ce3245
Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>