refactor(lib/attestation): handle platform token request in pieces
The size of the platform token requested from EL3 was limited by the
size of the shared buffer between RMM and the EL3 monitor. This patch
adds the necessary functionality in RMM for it to be able to request
the platform token in multiple requests, so the different pieces of
the token fit in the shared buffer. RMM now requests from EL3 the
pieces of the token sequentially and re-builds it on its side as it
receives the pieces.

Note: RMM now expect RMM-EL3 interface v0.3. Ensure that the EL3
firmware (TF-A) supports this version.

Change-Id: I9647ff2437e7fff57b80cb4c0ef20ac5bf495099
Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>