refactor(lib/attestation): manage token state within component
This patch refactors the `attestation` component to manage the
token generation state within itself. Hence the state transitions
are moved from the caller to the attestation APIs. An additional
state is added to indicate that the token sign context has been
initialized.

Overall the robustness of attestation state transitions are
improved and the callers of the attestation APIs are
simplified.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: Ie8e5d96b61f8603ce0ac1151433228eba712242b