refactor(plat/common): order mapped regions properly
At the moment we just list the regions we want to map in an array, but
later reference one particular one with its assumed index.
This is fragile and becomes more error prone when new regions are added.

Describe the different regions in an enum, and explicitly put each
region into its designated slot when initialising the array.
Then use this named index, when referencing the shared buffer region
later.

MISRA complains about the enum used for the index being signed, so
funnel this through an extra cast to make it happy.

Change-Id: I8926d7f04e6ae72c55b90a43d2f05aa4e2148b24
Signed-off-by: Andre Przywara <andre.przywara@arm.com>