fix(lib/measurement): fix misuse of psa_hash_finish()
do_extend() function in the measurement module invokes
psa_hash_finish(). However, parameter `hash_size` was set with the
value of PSA_HASH_LENGTH(psa_algorithm), which was not correct,
as the value has to be the size of the output buffer passed as
parameter `hash`. This patch corrects that behavior by adding extra
argument `out_size`, which contains the size of the output buffer.

Change-Id: Ide9b67f97dcf10a80f13cf705e5d4cbe7c640dfe
Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>