refactor(runtime): remove pointer of pmu field in ns_state
In order to simplify the handling of the PMU state in the ns_state
structure, the pointer has been removed and now the field is a
value field, so it does not need to be assigned anymore as it is
stored in the same ns_state structure. In addition, the logic for
handling this field was changed accordingly.

Change-Id: I697f9d76c463b2594a44ae7b3dc2ef9fffb0a852
Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>