feat(lib/attestation): add support for EL3 attestation queue
This patch adds code to enable the use of a EL3 service to queue
attestation requests. When ATTEST_EL3_TOKEN_SIGN is enabled,  the
added code initializes call backs to take the CBOR encoded hash
request to be enqueued and signed in EL3. The t_cose adapter invokes
the callbacks as part of signing process. The code added also
provides functions that can be called to pull responses from EL3.

Change-Id: Icbc1139f9c5fee5e1a229a964fdda291a332a8b8
Signed-off-by: Raghu Krishnamurthy <raghupathyk@nvidia.com>