feat(runtime/rmi): implement RMI_PDEV_STOP
RMI:
This RMI handler set the PDEV state to STOPPING, this allows NS Host to
invoke PDEV communicate call to do deinit the device connection and
transition the PDEV to STOPPED state.

DEV_ASSIGN App:
Added command to deinit the connection. This terminates the secure
session. This command also unset the public key context and certificate
chain hash in the libspdm connection state.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: I7d0d9b1654137f739793005cf290b8bb4c74957a