fix(lib/arch): have a default PA range
Currently the PA Range capability of the CPU is detected
from the corresponding encoding in id_aa64mmfr0_el1 register.
It is possible that RMM may run unmodified on a CPU which may
have an encoding higher than the known encodings to RMM.
In this case, this patch introduces a default PA range value
depending on if LPA2 is supported or not.

Change-Id: If9188c010bbc6bbffcbc3fdefa92451893e663ab
Signed-off-by: Soby Mathew <soby.mathew@arm.com>