feat(runtime): add support for lower and higher RMI/RSI revision
Add support in RMI_VERSION and RSI_VERSION commands to report lower
and higher supported interface revisions.

This allows the Host and the RMM or the Realm and the RMM to determine
whether there exists a mutually acceptable revision of the RMI or RSI
interface respectively via which the two components can communicate.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: I81bbe483543e43956cb5b4aa94c083f8bb930c44