refactor: import libspdm via git submodules
Previously the build system will fetch the libspdm
repo if RMM_V1_1 flag is set. This is now made part
of usual external dependency for RMM via git submodules.
Making this part of submodules unifies the dependency
management in RMM.

A later patch adds the subrepository dependency update
logic to the config phase of build.

Change-Id: Ib3539a1ac75960dfb05010bdc7334fe0a25d6a32
Signed-off-by: Soby Mathew <soby.mathew@arm.com>