fix(lib/rmm_el3_ifc): add console name to checksum calculation
The name field of console_info structure was missed
in checksum calculation. This is corrected by adding
a new helper checksum_calc() which computes the
checksum in a field agnostic manner.

Note that this change is a breaking change as the checksum
has to match the value from EL3. The corresponding
TF-A patch is merged here:

https://git.trustedfirmware.org/plugins/gitiles/TF-A/trusted-firmware-a/+/aa99881d30114e4e3a34eb4cc9808bac5f04226b%5E%21/

Change-Id: I2a766470028fbeff9acdd844564d1180d425b0e5
Signed-off-by: AlexeiFedorov <Alexei.Fedorov@arm.com>