feat(runtime/rmi): implement RMI_VDEV_COMMUNICATE
VDEV communicate is a RMI interface to do device communication for a
pending RDEV operation. The RDEV operation is set by RSI handler. This
model of Realm initiated communication flow will be changed to Host
initiated communication flow from alp13.

VDEV communicate uses the PDEV's communcation state as there is only one
connection (session) to the device exist and it is created and mananged
by PDEV.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: I374ac4f1e0efc5316043a17a12579dbc28fc73a6