fix(runtime/core): clear granules memory after mmu enabled
The size of the granules memory can vary significantly depending
on the platform. Currently, it is included as part of the .bss
segment, which is cleared before the MMU is enabled. Since memory
store/load operations are slower when the MMU is disabled, having
a larger .bss segment results in slower boot times. As the granules
memory is not used until much later in the boot process, separating
it from the .bss segment and clearing it only after the MMU is
enabled can improve boot-time performance.

Change-Id: Ie5b8a3e27a722f32fa2b8cf4bfe9b38f5962e561
Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>
Signed-off-by: Raghu Krishnamurthy <raghupathyk@nvidia.com>