refactor(app/attestatio): optimise cca token creation
This patches optimizes RMM core's acces to the cca attestation token
once it is created by the attestation app. The idea is that the physical
pages that are used for the heap area in the RMM app can be accessed
directly from RMM core, as they are mapped either in the .bss of the RMM
core or in the slot buffer as REC AUX pages.

Change-Id: I09fac7d5dae5b0610e23a6c2b4dae3bc266af7a5
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>