build(app): print RMM Core offset during build
Since apps are now packaged into RMM binary, for symbol debug
of RMM, the offset of RMM core from start of bin needs to
be taken into consideration. Having this print from build
log is useful for developers. Note that this info
is also present in <build>/Debug/bundle_app_out.txt
in case of a Debug build.

A sample print is given below:
"INFO: The offset of the RMM bin is 122880 (0x1e000) bytes"

Signed-off-by: Jacob Man Chun Yiu <jacobmanchun.yiu@arm.com>
Change-Id: Iaadb825f206c50accf408175560c54f7ceab765f