refactor(plat/host): split rmm-host-common lib
This commit splits the rmm host common library in two parts:
 * One can be used by rmm runtime and apps (host_el2_el0_cmn)
 * The other can only be used by rmm-runtime (host_harness)

rmm-host-el2-el0-cmn: This library contains functionality that can be
used by apps and rmm-runtime as well. This library was created from
rmm-host-support, and commonly used parts of the former rmm-host-common
are moved into it. It implements subset of  APIs from host_harness.h and
host_utils.h which are used in both apps and rmm-runtime.

rmm-host-harness: This library contains the parts from the former
rmm-host-common that can only be used by the rmm runtime running in EL2.

The lib component depends on rmm-platform but the rmm-runtime links the
lib and rmm-platform in the right order. The callbacks for el3 monitor
call is replaced with direct call to the host_harness functions.

For the host test build, in addition to the rmm-runtime dependency
mentioned above, the unit tests are also dependent on rmm-platform.

Change-Id: Ia261c5d02fd4431d43850fecb0e63df0acf3bd1b
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>
Signed-off-by: Soby Mathew <soby.mathew@arm.com>